moved visual effect functions into their own file
